+++ CHANGELOG.rst
@@ -1,6 +1,21 @@
 Change Log
 ==========
 
+v3.3.3
+----------
+- Add ``verbose`` option to more parsing/plotting functions for better control of output detail.
+- Improve effiency & robustness of oxidation state handling.
+- Miscellaneous efficiency (e.g. memory reduction) and robustness updates.
+- Improved GitHub Actions test efficiency.
+
+v3.3.2
+----------
+- Add ``verbose`` options to ``io.parse_energies()`` and ``snb-parse``, also used in ``snb-plot`` and
+  ``snb-analyse``, and set to ``False`` by default to reduce verbosity of certain SnB CLI commands.
+- Use ``doped`` functions to make oxi-state guessing (and thus defect initialisation) more efficient.
+- Miscellaneous efficiency and robustness updates.
+- Testing updates.
+
 v3.3.1
 ----------
 - ``distortion_metadata.json`` for each defect now saved to the individual defect folders (as well as the